POWERLINK IP core PDI re-design
---------------------
Detailed Description:
---------------------

1) Re-design of the control/status register of the PDI (meet 16bit alignment). Some additional features are added (e.g. event, led, asyncirq and FPGA revision number).
An asynchronous interrupt can be set by PCP. The irq is driven to the AP via an extra line (only valid for 8/16bit parallel interface and SPI) or it is or'd with the sync irq (valid for Avalon AP).
The asynchronous interrupt is controlled by the events set by the PCP or hardware. The PCP can set a generic event (GE) in combination with an argument and type field. The hardware event is generated by a phy link down.
The POWERLINK Error and Status LED is set by the PDI and can be accessed by the PCP and AP respectively. Further leds like phy link and activity are added and can be forced to certain values by the PCP and Ap respectively. The activity led is generated by using the MAC's TXEN and RXDV signals. It generates a 50Hz blink if activity is valid.
Two optinal leds are added and can be set by PCP and AP respectively.
Two asynchronous ping-pong buffers (TX/RX) are available.

2) clean up of src/pdi_par.vhd

3) openMAC Ethernet driver fix

HW-Changes:
-----------------
- led
* POWERLINK S / E
* PHY0 and 1 LINK
* PHY0 and 1 ACT
* two optional leds
* can be forced by PCP and AP (AP is the boss)
- event
* one generic event set by PCP with arg and type field
* PHY0 and 1 LINK DOWN event
* any asserted event generates asynchronous interrupt
* AP can poll or use irq
* PCP can observe if AP has acked the event
- asynchronous buffer
* two independent ping-pong buffers

SW-Changes:
-----------------
- within EdrvOpenMac.c
* tx msg buffer allocation was faulty for DPRAM TX packets


Status:
-----------------
openMAC verified with test application
PDI verified with test application

SPI, Direct I/O and 8/16bit parallel interface not yet verified

Compliance:
-----------------


------------------------------------------------- 
TODO:
--------
- hw acc
- automatic response delay compensation depending on connected phy
- 16bit SPI


git-svn-id: svn://ategge1109/fpga/POWERLINK/trunk@54 596af3a6-9f08-11df-8099-9b02302f7e83