[FIX] Fix openMAC descriptor memory mapping
The openMAC IP-Core provides up to 16 Rx and up to 16 Tx descriptors.
The Rx and Tx descriptors are implemented in separated memory blocks
configured to 512 kiB each. However, 16 descriptors only require the
half of each memory block - the other half is not used! Additionally,
the write accesses to the descriptors from the processor are duplicated
to both memory blocks unnecessarily. Anyhow, the MAC logic accesses the
descriptors from the second DPRAM ports correctly.
With this fix commit the descriptor memory blocks are reduced to 256 kiB
and the processor's write accesses are correctly directed to the
corresponding memory block.

Change-Id: I2b6ebc0389f99c1ffd374c270e4264521f8ad752